Split vector
